pal integration of face detect
