--- bt_nodes/hri/include/hri/stop_music.hpp
@@ -1,53 +1,0 @@
-// Copyright 2024 Intelligent Robotics Lab - Gentlebots
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions andGO2OBJECT
-// limitations under the License.
-
-#ifndef HRI__STOP_MUSIC_HPP_
-#define HRI__STOP_MUSIC_HPP_
-
-#include <string>
-
-#include "behaviortree_cpp_v3/behavior_tree.h"
-#include "behaviortree_cpp_v3/bt_factory.h"
-#include "hri/bt_service_node.hpp"
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_cascade_lifecycle/rclcpp_cascade_lifecycle.hpp"
-#include "std_srvs/srv/trigger.hpp"
-
-
-namespace hri
-{
-
-class StopMusic
-  : public hri::BtServiceNode<std_srvs::srv::Trigger,
-    rclcpp_cascade_lifecycle::CascadeLifecycleNode>
-{
-public:
-  explicit StopMusic(
-    const std::string & xml_tag_name, const std::string & action_name,
-    const BT::NodeConfiguration & conf);
-
-  void on_tick() override;
-  void on_result() override;
-
-  static BT::PortsList providedPorts()
-  {
-    return BT::PortsList(
-      {
-      });
-  }
-};
-
-}  // namespace hri
-
-#endif  // HRI__STOP_MUSIC_HPP_
