--- bt_nodes/hri/package.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>hri</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="samuele.sandrini@polito.it">Samuele</maintainer>
-  <license>Apache License, Version 2.0</license>
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  
-  <depend>rclcpp</depend>
-  <depend>rclcpp_cascade_lifecycle</depend>
-  <depend>rclcpp_action</depend>
-  <depend>behaviortree_cpp_v3</depend>
-  <depend>audio_common_msgs</depend>
-  <depend>whisper_msgs</depend>
-  <depend>llama_msgs</depend>
-  <depend>moveit_msgs</depend>
-  <depend>shape_msgs</depend>
-  <depend>std_msgs</depend>
-  <depend>gpsr_msgs</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
