+++ bt_nodes/robocup_hri/package.xml
@@ -1,34 +1,24 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>perception</name>
+  <name>robocup_hri</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="samuele.sandrini@polito.it">Samuele</maintainer>
-  <maintainer email="rodrigo.perez@urjc.es">rod</maintainer>
-  <maintainer email="juancarlos.serrano@urjc.es">Juanca</maintainer>
-  <maintainer email="alberto.gomezjacinto@urjc.es">Alberto</maintainer>
   <license>Apache License, Version 2.0</license>
-
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>ament_cmake_python</buildtool_depend>
-
+  
   <depend>rclcpp</depend>
   <depend>rclcpp_cascade_lifecycle</depend>
+  <depend>rclcpp_action</depend>
   <depend>behaviortree_cpp_v3</depend>
-  <depend>yolov8_msgs</depend>
+  <depend>audio_common_msgs</depend>
+  <depend>whisper_msgs</depend>
+  <depend>llama_msgs</depend>
   <depend>moveit_msgs</depend>
   <depend>shape_msgs</depend>
-  <depend>rclpy</depend>
-  <depend>tf2_ros</depend>
-  <depend>geometry_msgs</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>tf2</depend>
-  <depend>rclcpp_action</depend>
-  <depend>perception_system</depend>
-  <depend>perception_system_interfaces</depend>
-  <depend>sensor_msgs</depend>
-  <depend>hri</depend>
+  <depend>std_msgs</depend>
+  <depend>gpsr_msgs</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
